requirements.txt
bitsandbytes>=0.41.1
einops>=0.6.1
fire>=0.5.0
h5py>=3.8.0
loguru>=0.6.0
numpy>=1.19.5
packaging>=23.1
pandas>=1.2.0
pytorch_ignite==0.4.9
PyYAML==6.0.1
scikit_learn==1.3.0
SoundFile==0.10.3.post1
torch>=1.8.2
torch_audiomentations==0.11.0
torchaudio>=0.13.0
tqdm>=4.64.0
